Fix Selecting Future/Past Dates from Calendar
If you select a past or future date, the month is not decremented/incremented correctly.

This is because the value of `t.target` is for the nested div which contains the number, not for the div which has the date handler. When the `.is` test checks `t.target` for `.old` and `.new`, they are never present. Instead, use `t.currentTarget` to specify the element which is handling the event.